Rename nodes to expressions.
