+++ cpp/benchmarks/join/conditional_join_benchmark.cu
@@ -18,7 +18,7 @@
 
 #include <thrust/iterator/counting_iterator.h>
 
-#include <cudf/ast/nodes.hpp>
+#include <cudf/ast/expressions.hpp>
 #include <cudf/column/column_factories.hpp>
 #include <cudf/join.hpp>
 #include <cudf/table/table.hpp>
