+++ cpp/include/cudf/ast/expressions.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include <cudf/ast/detail/nodes.hpp>
+#include <cudf/ast/detail/expressions.hpp>
 #include <cudf/scalar/scalar.hpp>
 #include <cudf/scalar/scalar_device_view.cuh>
 #include <cudf/table/table_view.hpp>
