+++ cpp/src/ast/transform.cu
@@ -17,7 +17,7 @@
 #include <cudf/ast/detail/evaluator.cuh>
 #include <cudf/ast/detail/linearizer.hpp>
 #include <cudf/ast/detail/transform.cuh>
-#include <cudf/ast/nodes.hpp>
+#include <cudf/ast/expressions.hpp>
 #include <cudf/ast/transform.hpp>
 #include <cudf/column/column_device_view.cuh>
 #include <cudf/column/column_factories.hpp>
