+++ cpp/src/join/conditional_join.cu
@@ -16,7 +16,8 @@
 #include <join/conditional_join.cuh>
 #include <join/join_common_utils.hpp>
 
-#include <cudf/ast/nodes.hpp>
+#include <cudf/ast/expressions.hpp>
+//#include <cudf/detail/utilities/cuda.cuh>
 #include <cudf/join.hpp>
 #include <cudf/table/table.hpp>
 #include <cudf/table/table_view.hpp>
